Implemented output to a text file & restructed if statements dealing with argument parsing
I added the ability for the script to output to a text file. Due to this
added functionality I subsequently added a text file to the repository
currently called 'KnightcardSignInList.txt'. I also restructed the if
statements dealing with argument parsing because the if stmts checking
the number of arguments were incorectly used before and had no actual
effect on the script behavior.